Improve performance of the Interceptor evaluation for `items.xml`
